Move files and their useStyles to folders
