--- src/pages/lesson/LessonItem.module.css
@@ -1,17 +1,0 @@
-.lessonItem {
-  padding: 2px;
-  margin: 1px 0;
-  box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.2);
-  border-radius: 2px;
-  transition: 0.3s;
-  width: 40vw;
-  margin-top: 2vh;
-  margin-bottom: 2vh;
-  background-color: #fae1e4;
-}
-
-.lessonItemHover:hover {
-  box-shadow: 4px 4px 10px rgba(0, 0, 0, 0.3);
-  background-color: #ffc0cb;
-  cursor: pointer;
-}
