+++ src/pages/lesson/QuizItem/QuizItem.styles.ts
