+++ src/pages/lesson/QuizItem/QuizItem.tsx
@@ -6,8 +6,8 @@ import {
   Typography,
 } from "@mui/material";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import { QuizData } from "../../services/backend/types";
-import EditableActions from "../../components/EditableActions";
+import { QuizData } from "../../../services/backend/types";
+import EditableActions from "../../../components/EditableActions";
 import useStyles from "./QuizItem.styles";
 
 type QuizItemProps = {
